Conformed to pep8 standards
